requirements.txt
streamlit
python-docx
PyMuPDF
openai
python-dotenv
pytesseract
pdf2image
Pillow
crewai
langchain